블록체인 커넥터로 api전송

+++ src/main/java/com/ddiring/backend_market/api/blockchain/dto/trade/TradeDto.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class TradeDto {
-    private Integer tradeId;
+    private Long tradeId;
     private String projectId;
     private BuyInfoDto buyInfo;
     private SellInfoDto sellInfo;
